Switch datastore to MongoDB; add publish/audit/bundle, infra, docs
Datastore migration (backend-core, WS1+WS2):
- Replace EF Core + Npgsql with MongoDB.Driver 3.11.1. Delete CtmsDbContext,
  the design-time factory, all IEntityTypeConfiguration classes, the
  InitialCreate migration, and the dotnet-ef tool manifest.
- New Mongo persistence layer: CtmsMongoContext (typed collection accessors),
  MongoMappingRegistration (Guid _id as UUID v4, enums + Version as their
  natural BSON types, camelCase), repositories rewritten against collections,
  NoOpUnitOfWork, MongoHealthCheck (replaces the EF CanConnect readiness
  probe), MongoIndexInitializer + DataSeeder hosted services.
- Optimistic concurrency: TranslationString.Version uint/xmin -> app-managed
  long, guarded by a filtered UpdateAsync; mismatch still surfaces as 409
  with currentVersion. ConcurrencyException / DTO fields widened to long.
  Drop the EF DbUpdateConcurrencyException branch from the exception handler.

Lifecycle extensions:
- ReviewState gains Published; new "publish" action (Approved -> Published)
  and "reopen" now also covers Published -> NeedsReview. Editing a published
  or approved string returns it to NeedsReview.
- TranslationBundle aggregate: immutable (project, locale, version) snapshot
  of published strings with a SHA-256 content ETag. Repository + DTO added;
  HTTP endpoint deferred to WS4 (// TODO marker).
- AuditEntry aggregate + repository + AuditService: append-only trail written
  by TranslationStringService on every upsert and review transition. Indexes
  on (projectId, timestamp) and (entityType, entityId, timestamp).

Tests: moved from EF/SQLite-in-memory to EphemeralMongo (real mongod per
run, isolated database per test). 55 tests pass, build is warning-clean.

Containerisation + Azure infra (client-devops, WS8):
- Root Dockerfile (multi-stage, non-root, HTTP :8080), .dockerignore.
- docker-compose.yml: mongo:7 + redis:7-alpine + api with health gating;
  .env.example, override example. New config key ConnectionStrings:Redis.
- deploy/azure/main.bicep: ACR, user-assigned identity (AcrPull + KV Secrets
  User), Key Vault, Cosmos DB for MongoDB (ru/vcore switch), Azure Cache for
  Redis, Log Analytics, Container Apps env + API app with KV-reference
  secrets and /health probes. Compiles clean.

CI + docs (cicd-docs, WS9+WS11):
- azure-pipelines.yml + .azuredevops/templates: Build (restore/build/test
  with a mongo:7 service container) -> Package (publish + Docker buildAndPush
  to ACR) -> disabled Deploy stub referencing the Bicep.
- docs/: architecture.md (aggregates, lifecycle state diagram, bundle ETag,
  Mongo indexes), api.md, development.md, ADR 0001 + ADR 0002 (the
  Postgres -> MongoDB decision). docs/README.md indexes the set.

CLAUDE.md updated for MongoDB (persistence, index initializer + seeder in
place of migrations, data model, review table, EphemeralMongo tests).

+++ src/CTMS.Application/Common/ApplicationExceptions.cs
@@ -43,9 +43,9 @@ public sealed class ConflictException : Exception
 /// </summary>
 public sealed class ConcurrencyException : Exception
 {
-    public ConcurrencyException(uint currentVersion)
+    public ConcurrencyException(long currentVersion)
         : base("The translation string was changed by someone else; reload and retry.")
         => CurrentVersion = currentVersion;
 
-    public uint CurrentVersion { get; }
+    public long CurrentVersion { get; }
 }
